refactor: move filter traits to query traits namespace.
Ensuring that this tratis refers to query parameters traits
