+++ src/Items/Infrastructure/Http/Forms/QueryTraits/MapsName.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace AqHub\Items\Infrastructure\Http\Forms\Traits;
+namespace AqHub\Items\Infrastructure\Http\Forms\QueryTraits;
 
 use AqHub\Items\Domain\ValueObjects\Name;
 use Symfony\Component\HttpFoundation\Request;
